+++ src/main/scala/co/topl/brambl/cli/validation/SimpleTransactionValidationModule.scala
@@ -196,29 +196,12 @@ trait SimpleTransactionValidationModule {
     ).sequence.map(_ => paramConfig)
   }
 
-  def validateListParams(
-      paramConfig: BramblCliParams
-  ): ValidatedNel[String, BramblCliParams] = {
-    import cats.implicits._
-    List(
-      validateInputFile(
-        "Wallet DB",
-        paramConfig.someWalletFile,
-        required = true
-      )
-    ).sequence.map(_ => paramConfig)
-  }
 
   def validateAddEntitiyParams(
       paramConfig: BramblCliParams
   ): ValidatedNel[String, BramblCliParams] = {
     import cats.implicits._
     List(
-      validateInputFile(
-        "Wallet DB",
-        paramConfig.someWalletFile,
-        required = true
-      ),
       validateNonEmpty("Party name", paramConfig.partyName)
     ).sequence.map(_ => paramConfig)
   }
@@ -228,11 +211,6 @@ trait SimpleTransactionValidationModule {
   ): ValidatedNel[String, BramblCliParams] = {
     import cats.implicits._
     List(
-      validateInputFile(
-        "Wallet DB",
-        paramConfig.someWalletFile,
-        required = true
-      ),
       validateNonEmpty("Contract name", paramConfig.contractName),
       validateNonEmpty("Contract template", paramConfig.lockTemplate)
     ).sequence.map(_ => paramConfig)